+++ Source/Tests/EditorMAUI/MainPage.xaml.cs
@@ -4,6 +4,7 @@ using SkiaSharp;
 using SkiaSharp.Views.Maui;
 using SkiaSharp.Views.Maui.Controls;
 using Alternet.Editor;
+using Alternet.Drawing;
 
 namespace EditorMAUI;
 
@@ -75,13 +76,33 @@ public partial class MainPage : ContentPage
 
         editor.Interior.CornerClick += Interior_CornerClick;
         editor.Interior.Scroll += Interior_Scroll;
+        editor.Editor.LongTap += Editor_LongTap;
+
+        editor.Editor.CanLongTap = true;
+    }
+
+    private void Editor_LongTap(object? sender, Alternet.UI.LongTapEventArgs e)
+    {
+        LogToEntry("LongTap", e, true);
+        var caret = editor.Editor.CaretInfo;
+        if(caret is not null)
+        {
+            caret.OverlayColor = LightDarkColors.Blue;
+            caret.TopAndBottomOverlayVisible = true;
+            editor.Editor.InvalidateCaret();
+        }
     }
 
     private void Interior_Scroll(object? sender, Alternet.UI.ScrollEventArgs e)
     {
-        Alternet.UI.DebugUtils.DebugCallIf(false, () =>
+        LogToEntry("Scroll", e, false);
+    }
+
+    private void LogToEntry(string prefix, object obj, bool condition)
+    {
+        Alternet.UI.DebugUtils.DebugCallIf(condition, () =>
         {
-            entry1.Text = $"{Alternet.UI.LogUtils.GenNewId()} {e}";
+            entry1.Text = $"({Alternet.UI.LogUtils.GenNewId()}){prefix}: {obj}";
         });
     }