+++ whatsnew.md
@@ -1,3 +1,23 @@
+# 0.9.637 (2024 October 8)
+
+- Window: Fix exception on close if tooltip is shown.
+- Control.CaretInfo. Allows to customize internally painted caret.
+- Control.LongTap event and related methods.
+- MAUI: Fixed scrollbar movement on Android.
+- MAUI: Fixed scrollbar touch events handling.
+- MAUI: pan and double tap gestures support in ControlView.
+- Timer.TickAction, RichToolTip.HideDefault.
+- Control: VisualStatesOverride, VisualStates.
+- Control: RaiseScroll overload, RaiseScrollPageDown, RaiseScrollPageUp, RaiseScrollPageLeft, RaiseScrollPageRight.
+- BaseObject: StaticOptions, UseNamesInToString.
+- InteriorDrawable: ElementClick, Scroll events.
+- InteriorNotification.SendScrollToControl.
+- MAUI: RequireSwipeGesture, OnSwipeRight, OnSwipeLeft, OnSwipeUp, OnSwipeDown, OnSwipeGesture in ControlView.
+- DateUtils: Ticks to/from milliseconds.
+- KnownSvgImages: ImgCircleFilled, ImgDiamondFilled.
+- RectD, RectI: TopLineCenter, BottomLineCenter, set method for Center property.
+- SpeedColorButton: ShowColorSelector, ShowColorPopup, ShowColorDialog.
+
 # 0.9.636 (2024 October 4)
 
 - MenuItem: Opened, Closed, Highlighted events.